+++ andes/models/experimental/picontroller.py
@@ -1,0 +1,38 @@
+from andes.core import ModelData, NumParam, Model, State, Algeb, HardLimiter
+
+
+class PI2Data(ModelData):
+    """Data for PI2 model with deadlock issue"""
+
+    def __init__(self):
+        ModelData.__init__(self)
+        self.Kp = NumParam()
+        self.Ki = NumParam()
+        self.Wmax = NumParam()
+        self.Wmin = NumParam()
+
+
+class PI2Model(Model):
+    def __init__(self, system, config):
+        Model.__init__(self, system, config)
+        self.group = 'Experimental'
+        self.flags.update({'tds': True})
+        self.uin = State(v_str=0,
+                         e_str='Piecewise((0, dae_t<= 0), (1, dae_t <= 2), (-1, dae_t <6), (1, True))',
+                         tex_name='u_{in}',
+                         )
+        self.x = State(e_str='uin * Ki * HL_zi',
+                       v_str=0.05,
+                       )
+        self.y = Algeb(e_str='uin * Kp + x - y',
+                       v_str=0.05)
+
+        self.HL = HardLimiter(u=self.y, lower=self.Wmin, upper=self.Wmax)
+        self.w = Algeb(e_str='HL_zi * y + HL_zl * Wmin + HL_zu * Wmax - w',
+                       v_str=0.05)
+
+
+class PI2(PI2Data, PI2Model):
+    def __init__(self, system, config):
+        PI2Data.__init__(self)
+        PI2Model.__init__(self, system, config)
